backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "G-AI Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.1",
    "openai==1.10.0",
    "httpx==0.26.0",
    "google-generativeai==0.3.2",
]
